# Conflicts:
#	app/src/androidTest/java/com/futsch1/medtimer/MedicineStockTest.kt
#	app/src/androidTest/java/com/futsch1/medtimer/NotificationTest.kt
#	app/src/androidTest/java/com/futsch1/medtimer/ReminderTest.kt
#	app/src/main/java/com/futsch1/medtimer/Autostart.kt
#	app/src/main/java/com/futsch1/medtimer/MainActivity.kt
#	app/src/main/java/com/futsch1/medtimer/overview/ManualDose.kt
#	app/src/main/java/com/futsch1/medtimer/reminders/AlarmProcessor.kt
#	app/src/main/java/com/futsch1/medtimer/reminders/NotificationProcessor.kt
#	app/src/main/java/com/futsch1/medtimer/reminders/ReminderNotificationProcessor.kt
#	app/src/main/java/com/futsch1/medtimer/reminders/ReminderProcessorBroadcastReceiver.kt
#	app/src/main/java/com/futsch1/medtimer/reminders/RemoteInputReceiver.kt
#	app/src/main/java/com/futsch1/medtimer/reminders/RepeatProcessor.kt
#	app/src/main/java/com/futsch1/medtimer/reminders/SnoozeProcessor.kt
#	app/src/test/java/com/futsch1/medtimer/processortests/RepeatProcessorTest.kt

+++ app/src/main/java/com/futsch1/medtimer/reminders/SnoozeProcessor.kt
@@ -20,8 +20,8 @@ open class SnoozeProcessor @Inject constructor(
     val notificationProcessor: NotificationProcessor
 ) {
 
-    fun processSnooze(reminderNotificationData: ReminderNotificationData, snoozeTime: Int) {
-        reminderNotificationData.remindInstant = Instant.now().plusSeconds(snoozeTime * 60L)
+    fun processSnooze(reminderNotificationData: ReminderNotificationData, snoozeTime: Long) {
+        reminderNotificationData.remindInstant = Instant.now().plusSeconds(snoozeTime * 60)
         Log.d(LogTags.REMINDER, "Snoozing reminder: $reminderNotificationData")
 
         // Cancel a potential repeat alarm
